pyproject.toml
[project]
name = "microos"
version = "0.1.0"
description = "Atomic Microkernel Architecture optimized for AI-Driven Development"
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.22.1",
    "bcrypt==3.2.2",
    "fastapi>=0.115.0",
    "passlib>=1.7.4",
    "pydantic[email]>=2.10.0",
    "pyjwt>=2.9.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.30.0",
    "websockets>=13.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]
